a few bug fixed. still a lot of bug remains.
