+++ lib/app/modules/customer/customer_home/controllers/customer_home_controller.dart
@@ -1,11 +1,12 @@
-import 'package:get/get.dart';
 import 'package:dirham_uae/app/data/local/my_shared_pref.dart';
 import 'package:dirham_uae/app/modules/customer/customer_home/model/get_customer_service_model.dart';
 import 'package:dirham_uae/app/services/base_client.dart';
-import 'package:dirham_uae/utils/constants.dart';
+import 'package:dirham_uae/utils/urls.dart';
+import 'package:get/get.dart';
 
 class CustomerHomeController extends GetxController {
-  RxObjectMixin<GetCustomerServiceModel> getCustomerModel =GetCustomerServiceModel().obs;
+  RxObjectMixin<GetCustomerServiceModel> getCustomerModel =
+      GetCustomerServiceModel().obs;
   var isLoading = false.obs;
 
   getCustomeService() async {
@@ -17,10 +18,10 @@ class CustomerHomeController extends GetxController {
         'Authorization':
             'Bearer ${MySharedPref.getToken("token".obs).toString()}',
       },
-
       onSuccess: (response) {
         if (response.statusCode == 200) {
-          getCustomerModel.value = GetCustomerServiceModel.fromJson(response.data);
+          getCustomerModel.value =
+              GetCustomerServiceModel.fromJson(response.data);
           print(getCustomerModel.value.data!.data);
         }
         isLoading.value = false;
