+++ lib/app/modules/customer/customer_service_details/controllers/customer_service_details_controller.dart
@@ -1,33 +1,32 @@
-import 'package:get/get.dart';
 import 'package:dirham_uae/app/data/local/my_shared_pref.dart';
 import 'package:dirham_uae/app/modules/customer/customer_service_details/model/customer_service_details.dart';
 import 'package:dirham_uae/app/services/base_client.dart';
-import 'package:dirham_uae/utils/constants.dart';
+import 'package:dirham_uae/utils/urls.dart';
+import 'package:get/get.dart';
 
 class CustomerServiceDetailsController extends GetxController {
-  
+  var isLoading = false.obs;
 
- var isLoading= false.obs;
+  Rx<CustomerServiceDetailsModel> customerServiceDetailsModel =
+      CustomerServiceDetailsModel().obs;
 
- Rx<CustomerServiceDetailsModel> customerServiceDetailsModel= CustomerServiceDetailsModel().obs;
-
-  getServiceDetails(int id)async{
-    isLoading.value=true;
-  await BaseClient.safeApiCall(
-    "${Constants.ServiceCustomerDetails}/$id", 
-    RequestType.get,
-   headers: {
+  getServiceDetails(int id) async {
+    isLoading.value = true;
+    await BaseClient.safeApiCall(
+      "${Constants.ServiceCustomerDetails}/$id",
+      RequestType.get,
+      headers: {
         'Authorization':
             'Bearer ${MySharedPref.getToken("token".obs).toString()}',
       },
-
-    onSuccess:  (response) {
-      if (response.statusCode==200) {
-        customerServiceDetailsModel.value= CustomerServiceDetailsModel.fromJson(response.data);
-      }
-      isLoading.value=false;
-      update();
-    },
+      onSuccess: (response) {
+        if (response.statusCode == 200) {
+          customerServiceDetailsModel.value =
+              CustomerServiceDetailsModel.fromJson(response.data);
+        }
+        isLoading.value = false;
+        update();
+      },
     );
   }
 }
