+++ lib/app/modules/global/customer_tab/views/customertab_view.dart
@@ -1,6 +1,4 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:get/get.dart';
+import 'package:country_picker/country_picker.dart';
 import 'package:dirham_uae/app/components/custom_button.dart';
 import 'package:dirham_uae/app/components/custom_pass_field.dart';
 import 'package:dirham_uae/app/components/custom_snackbar.dart';
@@ -8,9 +6,12 @@ import 'package:dirham_uae/app/components/custom_text_field.dart';
 import 'package:dirham_uae/config/theme/light_theme_colors.dart';
 import 'package:dirham_uae/config/theme/my_images.dart';
 import 'package:dirham_uae/config/theme/my_styles.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:get/get.dart';
+
 import '../../../../../utils/global_variable/divider.dart';
 import '../controllers/customertab_controller.dart';
-import 'package:country_picker/country_picker.dart';
 
 class CustomertabView extends StatefulWidget {
   const CustomertabView({Key? key}) : super(key: key);
@@ -19,18 +20,14 @@ class CustomertabView extends StatefulWidget {
   State<CustomertabView> createState() => _CustomertabViewState();
 }
 
-
 class _CustomertabViewState extends State<CustomertabView> {
-
-  
   @override
   Widget build(BuildContext context) {
     CustomertabController controller = Get.put(CustomertabController());
-   
-       final category = (controller.categoryModel.value.data?? [])
+
+    final category = (controller.categoryModel.value.data ?? [])
         .firstWhereOrNull((element) => element.id == element.id);
 
-      
     return Scaffold(
       body: Container(
         decoration: BoxDecoration(
@@ -170,98 +167,70 @@ class _CustomertabViewState extends State<CustomertabView> {
                     );
                   },
                 ),
+
                 gapHeight(size: 20.0.h),
 
-                Text("Select Country",
+                Text("Select Service Category",
                     style:
                         kTitleTextstyle.copyWith(fontWeight: FontWeight.w500)),
                 gapHeight(size: 5),
-                CustomTextField(
-                  fillColor: Colors.white,
-                  textInputType: TextInputType.number,
-                  hintText: "Select country",
-                  controller: controller.country,
-                  readOnly: true,
-                  isIcon: false,
-                  icon: Img.uaeIcon,
-                  prefixIconScale: 3,
-                  onTap: () {
-                    showCountryPicker(
-                      countryListTheme:
-                          CountryListThemeData(backgroundColor: Colors.black),
-                      context: context,
-                      showPhoneCode: false, // Set to true to show phone codes
-                      onSelect: (Country country) {
-                        setState(() {
-                          controller.country.text = country.name;
-                          print(" country ${controller.country.text}");
-                        });
-                      },
-                    );
-                  },
-                ),
-                gapHeight(size: 20.0.h),
-
-                Text("Select Category",
-                    style:
-                        kTitleTextstyle.copyWith(fontWeight: FontWeight.w500)),
-                gapHeight(size: 5),
-                Container(
-                    padding: EdgeInsets.symmetric(horizontal: 10, vertical: 5),
-                    decoration: BoxDecoration(
-                      color: LightThemeColors.secounderyColor,
-                      borderRadius: BorderRadius.circular(8.r),
-                      boxShadow: [
-                        BoxShadow(
-                          color: LightThemeColors.shadowColor,
-                          offset: Offset(1, 1),
-                          blurRadius: 5,
-                        )
-                      ],
-                    ),
-                    child: Padding(
-                      padding: EdgeInsets.symmetric(horizontal: 15.0),
-                      child: Center(
-                        child: DropdownButton(
-                            dropdownColor:
-                                LightThemeColors.primaryColor.withOpacity(.5),
-                            alignment: Alignment.centerLeft,
-                            value: controller.categoryyIds.value.isEmpty
-                                ? null
-                                : controller.categoryyIds.value,
-                            isExpanded: true,
-                            underline: SizedBox(),
-                            hint: const Text("Choose Category"),
-                            items:
-                                (controller.categoryModel.value.data ??
-                                        [])
-                                    .map((mapValue) {
-                              return DropdownMenuItem(
-                                value: mapValue.name.toString(),
-                                child: Text(
-                                  mapValue.name.toString(),
-                                  style: TextStyle(
-                                      color: LightThemeColors.whiteColor),
-                                ),
-                              );
-                            }).toList(),
-                            onChanged: (newValue) {
-                              final selectedZone = (controller
-                                          .categoryModel.value.data ??
-                                      [])
-                                  .firstWhere((zone) =>
-                                      zone.name.toString() == newValue);
-
-                              // Check if the selected zone is not null and has an ID
-                              // ignore: unnecessary_null_comparison
-                              if (selectedZone != null &&
-                                  selectedZone.id != null) {
-                                controller.categoryyIds.value = newValue.toString();
-                              }
-                            }),
+                Obx(() => Container(
+                      padding:
+                          EdgeInsets.symmetric(horizontal: 10, vertical: 5),
+                      decoration: BoxDecoration(
+                        color: LightThemeColors.secounderyColor,
+                        borderRadius: BorderRadius.circular(8.r),
+                        boxShadow: [
+                          BoxShadow(
+                            color: LightThemeColors.shadowColor,
+                            offset: Offset(1, 1),
+                            blurRadius: 5,
+                          )
+                        ],
                       ),
-                    ),
-                  ),
+                      child: Padding(
+                        padding: EdgeInsets.symmetric(horizontal: 15.0),
+                        child: Center(
+                          child: DropdownButton(
+                              dropdownColor:
+                                  LightThemeColors.primaryColor.withOpacity(.8),
+                              alignment: Alignment.centerLeft,
+                              value: controller.categoryyIds.value.isEmpty
+                                  ? null
+                                  : controller.categoryyIds.value,
+                              isExpanded: true,
+                              underline: SizedBox(),
+                              hint: const Text("Choose Category"),
+                              items: (controller.categoryModel.value.data ?? [])
+                                  .map((mapValue) {
+                                return DropdownMenuItem(
+                                  value: mapValue.name.toString(),
+                                  child: Text(
+                                    mapValue.name.toString(),
+                                    style: TextStyle(
+                                        color: LightThemeColors.whiteColor),
+                                  ),
+                                );
+                              }).toList(),
+                              onChanged: (newValue) {
+                                final selectedZone =
+                                    (controller.categoryModel.value.data ?? [])
+                                        .firstWhere((zone) =>
+                                            zone.name.toString() == newValue);
+
+                                // Check if the selected zone is not null and has an ID
+                                // ignore: unnecessary_null_comparison
+                                if (selectedZone != null &&
+                                    selectedZone.id != null) {
+                                  controller.categoryyIds.value =
+                                      newValue.toString();
+                                  controller.categooryId =
+                                      selectedZone.id; // category id is now set
+                                }
+                              }),
+                        ),
+                      ),
+                    )),
                 gapHeight(size: 20),
                 Center(
                   child: Container(
