+++ lib/app/modules/global/login/views/login_view.dart
@@ -1,17 +1,19 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:get/get.dart';
-import 'package:dirham_uae/app/modules/provider/login/views/customer_login_view.dart';
 import 'package:dirham_uae/config/theme/light_theme_colors.dart';
 import 'package:dirham_uae/config/theme/my_images.dart';
 import 'package:dirham_uae/config/theme/my_styles.dart';
 import 'package:dirham_uae/utils/global_variable/my_scaffold_background.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+import 'package:get/get.dart';
+
 import '../../../../../utils/global_variable/divider.dart';
 import '../controllers/login_controller.dart';
+import 'customer_login_view.dart';
 import 'provider_login_view.dart';
 
 class LoginView extends GetView<LoginController> {
   const LoginView({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     // LoginController controller = Get.put(LoginController());
