+++ lib/app/modules/global/provider_tab/bindings/provider_tab_binding.dart
