+++ lib/app/modules/intro_screen/splash/bindings/splash_binding.dart
