+++ lib/app/modules/provider/account_details/controllers/account_details_controller.dart
@@ -1,14 +1,16 @@
-import 'package:get/get.dart';
 import 'package:dirham_uae/app/components/custom_snackbar.dart';
 import 'package:dirham_uae/app/modules/provider/account_details/models/get_provider_info_model.dart';
 import 'package:dirham_uae/app/services/base_client.dart';
-import 'package:dirham_uae/utils/constants.dart';
+import 'package:dirham_uae/utils/urls.dart';
+import 'package:get/get.dart';
+
 import '../../../../data/local/my_shared_pref.dart';
 
 class AccountDetailsController extends GetxController {
   RxBool isProviderInfoloading = false.obs;
   RxObjectMixin<GetProviderInfoModel> getProviderInfoModel =
       GetProviderInfoModel().obs;
+
 // **************** Get Provider Info **************
   Future getProviderInfo() async {
     isProviderInfoloading.value = true;
@@ -20,14 +22,14 @@ class AccountDetailsController extends GetxController {
       },
       Constants.getProviderInfo,
       RequestType.get,
-
       onSuccess: (response) {
         if (response.statusCode == 200) {
           print("get customer info status code 200");
           getProviderInfoModel.value =
               GetProviderInfoModel.fromJson(response.data);
 
-          print("hellooooo ::: ${getProviderInfoModel.value.data!.provider!.name.toString()}");
+          print(
+              "hellooooo ::: ${getProviderInfoModel.value.data!.provider!.name.toString()}");
           // CustomSnackBar.showCustomToast(
           //     message: response.data['message'].toString());
         }
@@ -44,6 +46,7 @@ class AccountDetailsController extends GetxController {
   }
 
   final count = 0.obs;
+
   @override
   void onInit() {
     getProviderInfo();
